% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeSymptomsDiagnostics.R
\docType{package}
\name{PhenotypeSymptoms-package}
\alias{PhenotypeSymptoms}
\alias{PhenotypeSymptoms-package}
\title{PhenotypeSymptoms: Generating Cohort Diagnostics for the OHDSI Phenotype Library}
\description{
Generating cohort diagnostics for the cohort definitions in the OHDSI Phenotype Library.
}
\keyword{internal}
